Make quibble work with cached quibbled things
This took a few weird steps:

1. At quibble-time, store the file that called `quibble` along with the
stubbing
2. At require-time, if the file that called `require` also called 
`quibble`, then skip the cache for whatever it's requiring. 


This way, we can skip the cache when we know the user is testing 
something, but continue to leverage the cache for transitive 
dependencies, modules, built-ins, etc. It should make the entire
enterprise far less error-prone, since a lot of modules naturally expect
to be singletons.