Add an additional test for an edge case
I'm concerned about this breaking current behavior
since it's possible for a half dozen files to match the same 
extensionless module path. It's important that specific claims win,
and that whatever we do hues to Node's built-in resolve algorithm.

I think this means the solution needs to be a fallback in the event of 
not finding a match, combined with using substack's resolve.sync package
for making sure we get the resolution rules right