fixes #468: td.replaceEsm() multiple replaces not working
The bug was a stupid one, and am actually not sure how it worked before!
Any `quibble.esm` should have updated the generation to ensure that fresh non-cache copies are loaded
once there is a change in mocking. But the generation was only incremented once, in the first `quibble.esm`.
This was fixed by moving the increment operation out of the `if`.