fix: enable running on Windows
`import` on Windows requires using the `file://` url.  This can be used
on all platforms for absolute paths.

`hackErrorStackToGetCallerFile()` returns a pathname which includes a
leading '/'.  This must be removed on Windows for `path.resolve` but
added back along with converting '\' to '/' for the import path.  In
other words, the code must handle multiple forms of paths.

Signed-off-by: Chris. Webster <chris@webstech.net>